Renamed some files, added a compliant grey for text color, removed some unused code and installed some plugins that were missing. Also, some styling.
