+++ gatsby-browser.js
@@ -1,5 +1,1 @@
-// custom typefaces
-import "typeface-montserrat"
-import "typeface-merriweather"
-
 import "prismjs/themes/prism.css"
