+++ package.json
@@ -3,11 +3,12 @@
   "private": true,
   "description": "Homepage of Turku.py",
   "repository": {
-    "type" : "git",
-    "url" : "https://github.com/turkupy/turkupy.github.io/"
+    "type": "git",
+    "url": "https://github.com/turkupy/turkupy.github.io/"
   },
   "version": "0.1.0",
   "dependencies": {
+    "babel-plugin-styled-components": "^1.10.7",
     "gatsby": "^2.21.22",
     "gatsby-image": "^2.4.3",
     "gatsby-plugin-feed": "^2.5.1",
@@ -16,6 +17,7 @@
     "gatsby-plugin-offline": "^3.2.1",
     "gatsby-plugin-react-helmet": "^3.3.1",
     "gatsby-plugin-sharp": "^2.6.2",
+    "gatsby-plugin-styled-components": "^3.3.1",
     "gatsby-plugin-typography": "^2.5.1",
     "gatsby-remark-copy-linked-files": "^2.3.2",
     "gatsby-remark-images": "^3.3.2",
@@ -32,8 +34,6 @@
     "react-helmet": "^5.2.1",
     "react-typography": "^0.16.19",
     "styled-components": "^5.1.0",
-    "typeface-merriweather": "0.0.72",
-    "typeface-montserrat": "0.0.75",
     "typography": "^0.16.19",
     "typography-theme-grand-view": "^0.16.19",
     "typography-theme-wordpress-2016": "^0.16.19"
