+++ src/components/Layout.jsx
@@ -1,8 +1,18 @@
 import React from "react"
 import { Link } from "gatsby"
 import styled from "styled-components"
-import { rhythm, scale } from "../utils/typography"
 import logo from "../../content/assets/logo.png"
+import { createGlobalStyle } from "styled-components"
+import { colors } from '../utils/styles'
+
+const GlobalStyles = createGlobalStyle`
+  color: ${colors.darkGrey};
+
+  h1, h2 {
+    color: ${colors.darkGrey};
+  }
+`
+
 
 const StyledLayout = styled.div`
   display: flex;
@@ -24,15 +34,18 @@ const Footer = styled.footer`
 
 `
 
-const Layout = ({ location, title, children }) => {
+const Layout = ({ children }) => {
   return (
+    <>
+    <GlobalStyles />
     <StyledLayout>
-      <header><Link to ="/"><img src={logo} alt="Turku.py" /></Link></header>
+      <header><Link to ="/"><img src={logo} alt="Turku.py - programming community for women and nonbinaries" /></Link></header>
       <main>{children}</main>
       <Footer>
 
       </Footer>
     </StyledLayout>
+    </>
   )
 }
