+++ src/templates/EventPage.jsx
@@ -2,21 +2,23 @@ import React from 'react'
 import Layout from '../components/Layout'
 import styled from 'styled-components'
 import { colors } from "../utils/styles"
+import Seo from "../components/Seo"
 
 const EventPage = ({ data, location }) => {
   const event = data.markdownRemark
 
 
-const EventTitle = styled.h2`
+const EventTitle = styled.h1`
   margin-bottom: .4em;
 `
 
 const EventDetail = styled.p`
-  color: ${colors.grey};
+  color: ${colors.darkGrey};
   padding-bottom: 1.4em;
 `
 
 return <Layout>
+  <Seo title={data.markdownRemark.frontmatter.title} />
   <EventTitle>{data.markdownRemark.frontmatter.title}</EventTitle>
   <EventDetail>{data.markdownRemark.frontmatter.date} @ {data.markdownRemark.frontmatter.host}</EventDetail>
 
@@ -28,7 +30,7 @@ export default EventPage
 
 
 export const pageQuery = graphql`
-  query BlogPostBySlug($slug: String!) {
+  query EventsBySlug($slug: String!) {
     site {
       siteMetadata {
         title
