+++ src/utils/styles.js
@@ -2,5 +2,6 @@ export const colors = {
   lightPink: '#f06b8e',
   darkPink: '#ed2859',
   grey: '#808285',
+  darkGrey: '#4D4E51',
   white: '#FFFFFF'
 }
